+++ examples/plot_pt_vs_eff.py
@@ -7,7 +7,7 @@ from puma.utils import get_dummy_2_taggers, logger
 
 # define a small function to calculate discriminant
 def disc_fct(arr: np.ndarray, f_c: float = 0.018) -> np.ndarray:
-    """Tagger discriminant
+    """Tagger discriminant.
 
     Parameters
     ----------