+++ puma/plot_base.py
@@ -14,7 +14,7 @@ atlasify.LINE_SPACING = 1.3  # overwrite the default, which is 1.2
 # TODO: enable `kw_only` when switching to Python 3.10
 # @dataclass(kw_only=True)
 @dataclass
-class PlotLineObject:  # pylint: disable=too-many-instance-attributes
+class PlotLineObject:
     """Base data class defining properties of a plot object.
 
     Parameters
@@ -59,7 +59,7 @@ class PlotLineObject:  # pylint: disable=too-many-instance-attributes
 # TODO: enable `kw_only` when switching to Python 3.10
 # @dataclass(kw_only=True)
 @dataclass
-class PlotObject:  # pylint: disable=too-many-instance-attributes
+class PlotObject:
     """Data base class defining properties of a plot object.
 
     Parameters
@@ -232,7 +232,7 @@ class PlotObject:  # pylint: disable=too-many-instance-attributes
             )
 
     def __check_figsize(self):
-        """Check `figsize`
+        """Check `figsize`.
 
         Raises
         ------
@@ -250,7 +250,7 @@ class PlotObject:  # pylint: disable=too-many-instance-attributes
             )
 
     def __check_yratio(self, yratio):
-        """Check `yratio`
+        """Check `yratio`.
 
         Parameters
         ----------
@@ -271,11 +271,11 @@ class PlotObject:  # pylint: disable=too-many-instance-attributes
             )
 
 
-class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
-    """Base class for plotting"""
+class PlotBase(PlotObject):
+    """Base class for plotting."""
 
     def __init__(self, **kwargs) -> None:
-        """Initialise class
+        """Initialise class.
 
         Parameters
         ----------
@@ -351,8 +351,8 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         same_height: bool = False,
         colour: str = "#000000",
         fontsize: int = 10,
-    ):  # pylint: disable=too-many-arguments
-        """Drawing working points in plot
+    ):
+        """Drawing working points in plot.
 
         Parameters
         ----------
@@ -431,7 +431,6 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         force_y : bool, optional
             Forcing log on y-axis even if `logy` attribute is False, by default False
         """
-
         if self.logx or force_x:
             if not self.logx:
                 logger.warning(
@@ -543,7 +542,7 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
                 ratio_axis.tick_params(axis="x", labelsize=labelsize, **kwargs)
 
     def set_xlim(self, xmin: float = None, xmax: float = None, **kwargs):
-        """Set limits of x-axis
+        """Set limits of x-axis.
 
         Parameters
         ----------
@@ -589,7 +588,7 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         )
 
     def atlasify(self, use_tag: bool = True, force: bool = False):
-        """Apply ATLAS style to all axes using the atlasify package
+        """Apply ATLAS style to all axes using the atlasify package.
 
         Parameters
         ----------
@@ -599,7 +598,6 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         force : bool, optional
             Force ATLAS style also if class variable is False, by default False
         """
-
         if self.plotting_done is False and force is False:
             logger.warning(
                 "`atlasify()` has to be called after plotting --> "
@@ -614,7 +612,7 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
                 # TODO: for some reason, pylint complains about the used arguments
                 # when calling atlasify ("unexpected-keyword-arg") error
                 # --> fix this
-                atlasify.atlasify(  # pylint: disable=E1123
+                atlasify.atlasify(
                     atlas=self.atlas_first_tag,
                     subtext=self.atlas_second_tag,
                     axes=self.axis_top,
@@ -637,8 +635,8 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
             if force:
                 if self.apply_atlas_style is False:
                     logger.warning(
-                        "Initialising ATLAS style even though `apply_atlas_style` is  "
-                        "set to False."
+                        "Initialising ATLAS style even though `apply_atlas_style` is "
+                        " set to False."
                     )
                 if self.plotting_done is False:
                     logger.warning(
@@ -686,7 +684,7 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         loc: str = None,
         bbox_to_anchor: tuple = None,
         axis_for_legend=None,
-    ):  # pylint: disable=too-many-arguments
+    ):
         """Create a legend to indicate what different linestyles correspond to.
 
         Parameters
@@ -704,7 +702,6 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         axis_for_legend : matplotlib.Axes.axis, optional
             Axis on which to draw the legend, by default None
         """
-
         if axis_for_legend is None:
             axis_for_legend = self.axis_top
 
@@ -731,7 +728,7 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         axis_for_legend.add_artist(linestyle_legend)
 
     def set_ratio_label(self, ratio_panel: int, label: str):
-        """Associate the rejection class to a ratio panel
+        """Associate the rejection class to a ratio panel.
 
         Parameters
         ----------
@@ -753,7 +750,7 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
         self.ylabel_ratio[ratio_panel - 1] = label
 
     def initialise_plot(self):
-        """Calls other methods which are usually used when plotting"""
+        """Calls other methods which are usually used when plotting."""
         self.set_title()
         self.set_log()
         self.set_y_lim()