+++ puma/tests/test_var_vs_eff.py
@@ -1,9 +1,7 @@
 #!/usr/bin/env python
 
 
-"""
-Unit test script for the functions in var_vs_eff.py
-"""
+"""Unit test script for the functions in var_vs_eff.py."""
 
 import os
 import tempfile
@@ -219,10 +217,8 @@ class VarVsEffTestCase(unittest.TestCase):
             var_plot.divide(var_plot_comp, mode="sig_eff")
 
 
-class VarVsEffOutputTestCase(
-    unittest.TestCase
-):  # pylint:disable=too-many-instance-attributes
-    """Test class for the puma.var_vs_eff_plot output"""
+class VarVsEffOutputTestCase(unittest.TestCase):
+    """Test class for the puma.var_vs_eff_plot output."""
 
     def setUp(self):
         # Set up temp directory for comparison plots