+++ puma/tests/utils/test_generate.py
@@ -1,11 +1,10 @@
 #!/usr/bin/env python
-"""
-Unit test script for the functions in utils/generate.py
-"""
+"""Unit test script for the functions in utils/generate.py."""
 
 import unittest
 
 import numpy as np
+from numpy.lib.recfunctions import structured_to_unstructured as s2u
 
 from puma.utils import logger, set_log_level
 from puma.utils.generate import get_dummy_2_taggers, get_dummy_multiclass_scores
@@ -27,12 +26,12 @@ class GetDummyMulticlassScoresTestCase(unittest.TestCase):
             self.assertEqual(len(labels), 9)
 
     def test_range(self):
-        """Check that correct range of output is returned"""
+        """Check that correct range of output is returned."""
         output, _ = get_dummy_multiclass_scores()
         with self.subTest("max val"):
-            self.assertLessEqual(np.max(output), 1)
+            self.assertLessEqual(np.max(s2u(output)), 1)
         with self.subTest("min val"):
-            self.assertGreaterEqual(np.min(output), 0)
+            self.assertGreaterEqual(np.min(s2u(output)), 0)
 
 
 class GetDummy2TaggersTestCase(unittest.TestCase):