+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### [Latest]
 
+- Update pre-commit (using ruff) and store tagger scores as structured array [!177](https://github.com/umami-hep/puma/pull/177)
 - Remove dev image [!176](https://github.com/umami-hep/puma/pull/176)
 - Fix bug in ratio axis limits [!175](https://github.com/umami-hep/puma/pull/175)
 - Add `VarVsVar` plot [!172](https://github.com/umami-hep/puma/pull/172)